requirements.txt
asgiref==3.11.0
boto3==1.42.21
botocore==1.42.21
Django==5.1.4
django-cors-headers==4.9.0
django-filter
django-storages==1.14.6
djangorestframework==3.16.1
drf-yasg==1.21.11
gunicorn==23.0.0
inflection==0.5.1
jmespath==1.0.1
Markdown==3.10
packaging==25.0
pillow==12.1.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
uritemplate==4.2.0
urllib3==2.6.2
